task 0

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	char word[] = "_putchar";
 	int i = 0;
-	while (i < sizeof(word) - 1)
+	while (i < (int)sizeof(word) - 1)
 	{
 		_putchar(word[i]);
 		i++;
